% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_geo_lines_plot}
\alias{data_geo_lines_plot}
\title{Geolines data}
\format{A data frame with 617 rows and 9 variables:
\describe{
  \item{sender.location}{Country, City}
  \item{start.latitude}{start latitude}
  \item{start.longitude}{start longitude}
  \item{end.latitude}{end latitude}
  \item{end.longitude}{end longitude}
  \item{end.location}{Country, City}
  \item{date}{YYYY-MM-DD}
  \item{journey}{Concatenated start - end coordinate}
  \item{number.of.letters}{Number of letters sent across route}
  ...
}}
\source{
\url{http://www.idn.it.ox.ac.uk}
}
\usage{
data_geo_lines_plot
}
\description{
A (long-formatted) dataset designed to demonstrate different options for geoline charts (great circles between points on a map).
The dataset has been based on an academic dataset that will be cited after the embargo period has come to an end.
}
\keyword{datasets}
