% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_line_chart}
\alias{data_line_chart}
\title{Line chart data}
\format{A data frame with 35 rows and 4 variables:
\describe{
  \item{series}{Different series (analagous to different traces/lines in the chart)}
  \item{af}{x-values in sample charts}
  \item{fdr}{y-values in sample charts}
  \item{tpr}{y-values in sample charts}
  \item{color}{color for the trace/line}
}}
\source{
\url{https://dx.doi.org/10.6084/m9.figshare.4555441.v1}
}
\usage{
data_line_chart
}
\description{
A (long-formatted) dataset designed to demonstrate different options for line charts.
}
\keyword{datasets}
