% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_scatter_plot}
\alias{data_scatter_plot}
\title{Scatter chart data}
\format{A data frame with 100 rows and 4 variables:
\describe{
  \item{x}{x-values in sample charts}
  \item{y}{y-values in sample charts}
  \item{group}{Groups of points, analagous to different traces in the chart}
  \item{color}{color for the trace/point}
}}
\source{
\url{https://dx.doi.org/10.6084/m9.figshare.4555441.v1}
}
\usage{
data_scatter_plot
}
\description{
A (long-formatted) dataset designed to demonstrate different options for scatter charts.
}
\keyword{datasets}

