% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_stacked_bar_chart}
\alias{data_stacked_bar_chart}
\title{Stacked bar chart data}
\format{A data frame with 10 rows and 6 variables:
\describe{
  \item{country}{country where activity took place}
  \item{activity}{type of activity}
  \item{hours}{hours spent performing activity}
}}
\source{
\url{http://www.diamondse.info/}
}
\usage{
data_stacked_bar_chart
}
\description{
A (long-formatted) dataset designed to demonstrate different options for stacked bar charts.
}
\keyword{datasets}

