% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fancy_charts_nonstandard_eval.R
\name{fancy_chart}
\alias{fancy_chart}
\title{Learning non-standard evaluation}
\usage{
fancy_chart(data = NA, library = "plotly", x.axis = NA, y.axis = NA,
  ...)
}
\description{
\code{fancy_chart} is not to be used, it is only a place to learn non-standard evaluation.
}

