% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charts.R
\name{hc_stacked_bar_chart}
\alias{hc_stacked_bar_chart}
\title{highchart bar chart}
\usage{
hc_stacked_bar_chart(data = NA, categories.column = NA,
  categories.order = NA, subcategories.column = NA, value.column = NA,
  stacking.type = NA, subcategories.order = NA)
}
\description{
highchart bar chart
}

