% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-lines-plot.R
\name{journey_termini_data}
\alias{journey_termini_data}
\title{\code{journey_termini_data} should not be used directly.}
\usage{
journey_termini_data(data)
}
\description{
\code{journey_termini_data} should not be used directly.
}
