% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeseries.R
\name{calendar_heatmap}
\alias{calendar_heatmap}
\alias{rcharts_calendar_map}
\title{Calendar Heatmap.}
\usage{
calendar_heatmap(data, library = "googleVis")

rcharts_calendar_map(...)
}
\arguments{
\item{data}{Date to be included in the heatmap}

\item{library}{Which library to use, googleViz is default.}

\item{...}{all arguments other than \code{data} and \code{library} provided to \code{calendar_heatmap}.}
}
\value{
Stacked bar chart of the class specified by the specified library.
}
\description{
\code{rcharts_calendar_map} should not be used directly, it generates a calendar heatmap using rCharts.
}
\section{Warning}{
 These should be considered utility functions only, the magic is contained within the htmlwidget library,
 it is perfectly possible breaking changes to the library and/or the underling JavaScript library will
 break these. In such case, refer to http://martinjhnhadley.github.io/OxfordIDN_htmlwidgets for generic tutorials on
the widely used htmlwidget libraries and how to select an alternative visualisation library.
}

