% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-bubble-chart.R
\name{geo_bubble_chart}
\alias{geo_bubble_chart}
\alias{leaflet_geo_bubble_chart}
\alias{hc_geo_bubble_chart}
\alias{plotly_geo_bubble_chart}
\title{Geo bubble chart}
\usage{
geo_bubble_chart(data = NA, library = "leaflet", bubble.radius = ~radius,
  bubble.group, bubble.opacity = 0.6, color = ~color, popup.text = NULL,
  mouseover.text = NULL)

leaflet_geo_bubble_chart(...)

hc_geo_bubble_chart(...)

plotly_geo_bubble_chart(...)
}
\arguments{
\item{data}{A dataframe, must be long-formatted.}

\item{library}{Which library to use, highchart is default.}

\item{bubble.radius}{Column containing the bubble radius for each point. Must be given as formula, i.e. ~radius}

\item{popup.text}{Content to display when a marker is clicked. Must be given as formula, i.e. ~country}

\item{mouseover.text}{Content to display when the cursor passes over a marker. Must be given as formula, i.e. ~country}

\item{...}{all arguments provided to \code{geo_bubble_chart}.}

\item{...}{all arguments provided to \code{geo_bubble_chart}.}

\item{...}{all arguments provided to \code{geo_bubble_chart}.}
}
\description{
\code{geo_bubble_chart} creates a geo bubble chart using the specified library, which can be used in the library's \%>\% workflow. Data must be provided in long format.

\code{leaflet_geo_bubble_chart} should not be used directly, it generates a geo bubble plot using leaflet.

\code{hc_geo_bubble_chart} should not be used directly, it generates a geo bubble plot using highcharter.

\code{plotly_geo_bubble_chart} should not be used directly, it generates a geo bubble plot using plotly
}
