% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-choropleth-chart.R
\name{geo_choropleth_chart}
\alias{geo_choropleth_chart}
\alias{leaflet_geo_choropleth_chart}
\title{geo_choropleth_chart}
\usage{
geo_choropleth_chart(spdf = NA, library = "leaflet",
  region.border.width = 1, region.border.color = "#000080",
  region.value = NA, legend = list(type = "continuous", rcolorbrewer.palette
  = , minColor = NULL))

leaflet_geo_choropleth_chart(...)
}
\arguments{
\item{spdf}{A SpatialPolygonsDataFrame containing the data to be displayed}

\item{library}{Which library to use, leaflet is default.}

\item{region.border.width}{Width of region borders, default to 1}

\item{region.border.color}{Color of region borders, default to #000080}

\item{region.value}{Column within spdf@data containing the numerical , i.e sub-categorisations. Must be given as formula, i.e. ~country}

\item{...}{all arguments other than \code{data} and \code{library} provided to \code{geo_choropleth_chart}.}
}
\description{
\code{geo_choropleth_chart} creates a map comprised of shaded regions (where colour is used to allow visual comparison betweem region) using the specified library, which can be used in the library's \%>\% workflow. Data must be provided in long format.

\code{leaflet_geo_choropleth_chart} should not be used directly, it generates a map with great circles between points using Leaflet.
}
