% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-lines-plot.R
\name{geo_lines_plot}
\alias{geo_lines_plot}
\alias{leaflet_geo_lines_plot}
\title{geo_lines_plot}
\usage{
geo_lines_plot(data = NA, library = "leaflet", line.color = "#2c7bb6",
  line.popup = NULL, line.mouseover = NULL, start.color = "#fdae61",
  end.color = "#d7191c", both.color = "#7570b3",
  line.options = list(weight = 4), termini.legend = TRUE,
  termini.options = list(termini.radius = 4))

leaflet_geo_lines_plot(...)
}
\arguments{
\item{data}{A data.frame with start-end pairs, needs the following columns
\itemize{
 \item{"start.longitude"}{ : start longitude}
 \item{"start.latitude"}{ : start latitude}
 \item{"end.longitude"}{ : end longitude}
 \item{"end.latitude"}{ : end latitude}
 }}

\item{library}{must be one of the supported libraries, currently; "leaflet". Defaults to "leaflet".}

\item{line.color}{color for the geolines (great circles), defaults to #2c7bb6}

\item{line.popup}{expression used to populate popups displayed when a geoline is clicked, can use models/formula. For instance, ~paste(start.longitude, end.longitude)}

\item{start.color}{color of the dots representing start locations, defaults to #fdae61}

\item{end.color}{color of the dots representing end locations, defaults to #d7191c}

\item{line.options}{named list of options for geolines}

\item{termini.legend}{include a legend for termini, using \code{termini.options} colors. Default to TRUE.}

\item{termini.options}{named list of options for the termini (start/end dots)
\itemize{
  \item{start.fill}{ : Fill start termini? TRUE or FALSE, default FALSE.}
  \item{end.fill}{ : Fill end termini? TRUE or FALSE, default TRUE.}
  \item{termini.radius}{ : size of start termini.}
  \item{termini.radius}{ : size of start termini.}
 }}

\item{...}{all arguments other than \code{data} and \code{library} provided to \code{geo_lines_plot}.}
}
\description{
\code{geo_lines_plot} creates a map with lines (great circles) between geo graphic locations using the specified library, which can be used in the library's \%>\% workflow. Data must be provided in long format.

\code{leaflet_geo_lines_plot} should not be used directly, it generates a map with great circles between points using Leaflet.
}
