% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-marker-plot.R
\name{geo_marker_plot}
\alias{geo_marker_plot}
\alias{leaflet_geo_marker_plot}
\title{Geo marker plot}
\usage{
geo_marker_plot(data = NA, library = "leaflet", popup.text = NULL,
  mouseover.text = NULL)

leaflet_geo_marker_plot(...)
}
\arguments{
\item{data}{A dataframe, must be long-formatted.}

\item{library}{Which library to use, highchart is default.}

\item{popup.text}{Content to display when a marker is clicked. Must be given as formula, i.e. ~country}

\item{mouseover.text}{Content to display when the cursor passes over a marker. Must be given as formula, i.e. ~country}

\item{...}{all arguments provided to \code{geo_marker_plot}.}
}
\description{
\code{geo_marker_plot} creates a geo scatter plot using the specified library, which can be used in the library's \%>\% workflow. Data must be provided in long format.

\code{leaflet_stacked_bar_chart} should not be used directly, it generates a geo marker plot using leaflet.
}
