% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line-chart.R
\name{line_chart}
\alias{hc_line_chart}
\alias{line_chart}
\title{Line Charts.}
\usage{
line_chart(data = NA, library = "highcharter", x.column, y.column,
  traces.column, markers = TRUE)

hc_line_chart(...)
}
\arguments{
\item{data}{A dataframe, must be long-formatted.}

\item{library}{Which library to use, highchart is default.}

\item{x.column}{Column containing x-coordinates for data points}

\item{y.column}{Column containing y-cordinates for data points}

\item{traces.column}{Column containing the traces/trace names}

\item{markers}{Show plot markers on lines? TRUE or FALSE.}
}
\description{
\code{stacked_bar_chart} creates a stacked bar chart using the specified library, which can be used in the library's \%>\% workflow. Data must be provided in long format.

\code{hc_line_chart} should not be used directly, it generates a stacked barchart using Plotly.
}

