% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter-plot.R
\name{scatter_plot}
\alias{hc_scatter_plot}
\alias{scatter_plot}
\title{Scatter plot.}
\usage{
scatter_plot(data = NA, library = "highcharter", x.column, y.column,
  marker.size = 1, fillOpacity = 0.3, traces.column)

hc_scatter_plot(...)
}
\arguments{
\item{data}{A dataframe, must be long-formatted.}

\item{library}{Which library to use, highchart is default.}

\item{x.column}{Column containing x-coordinates for data points}

\item{y.column}{Column containing y-cordinates for data points}

\item{marker.size}{Size of markers (circles, by default)}

\item{fillOpacity}{Opacity of markers}

\item{traces.column}{Column containing the traces/trace names}
}
\description{
\code{scatter_plot} creates a scatter plot using the specified library, which can be used in the library's \%>\% workflow. Data must be provided in long format.

\code{hc_scatter_plot} should not be used directly, it generates a scatter plot using highcharter.
}

