% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oidnChaRts.R
\docType{package}
\name{oidnChaRts}
\alias{oidnChaRts}
\alias{oidnChaRts-package}
\title{oidnChaRts: A package  with utility functions for interactive charts.}
\description{
The package contains functions for creating the following types of data visualisations:
Charts, Maps, Networks
}
\section{Chart functions}{

stacked_bar_chart
}

\section{Maps functions}{

maps
}

\section{Networks}{

networks
}

