% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charts.R
\name{plotly_stacked_bar_chart}
\alias{plotly_stacked_bar_chart}
\title{Plotly Stacked Bar Chart.}
\usage{
plotly_stacked_bar_chart(data = NA, categories.column = NA,
  categories.order = NA, subcategories.column = NA, value.column = NA,
  stacking.type = NA, subcategories.order = NA)
}
\description{
Plotly Stacked Bar Chart.
}

