% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charts.R
\name{stacked_bar_chart}
\alias{stacked_bar_chart}
\title{Stacked Bar Chart.}
\usage{
stacked_bar_chart(data = NA, library = "highcharter",
  categories_column = NA, measure_columns = NA, stacking_type = NA,
  ordering_function = c, explicit_order = NA)
}
\arguments{
\item{data}{A dataframe.}

\item{library}{Which library to use, highchart is default.}

\item{categories_column}{Divisions within each bar.}

\item{measure_columns}{Column names containing data for bars (each category is a specific bar),
this could also be considered as the "y-axis" in a horizontally orientated bar chart.}

\item{stacking_type}{Stacking to apply to bars, default NULL. Possible values, c("percent", "normal")}

\item{ordering_function}{Function applied to bar segments to decide on order, default c.
It is a good default to choose var.}

\item{explicit_order}{Overrides ordering_function, order must be given by index.}
}
\value{
Stacked bar chart of the class specified by the specified library.
}
\description{
\code{stacked_bar_chart} creates a stacked bar chart using the specified library, which can be used in the libraries \%>\% workflow..
}
\examples{
stacked_bar_chart(1, 1)
stacked_bar_chart(1, "foo")

#' @section Warning:
These should be considered utility functions only, the magic is contained within the htmlwidget library,
it is perfectly possible breaking changes to the library and/or the underling JavaScript library will
break these. In such case, refer to http://ox-it.github.io/OxfordIDN_htmlwidgets for generic tutorials on
the widely used htmlwidget libraries and how to select an alternative visualisation library.
}

