% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_geo_marker_plot}
\alias{data_geo_marker_plot}
\title{Geo Marker plot data}
\format{A data frame with 157 rows and 5 variables:
\describe{
  \item{country}{Country in which point resides}
  \item{city}{City in which point resides, with state name in brackets where appropriate}
  \item{latitude}{latitude}
  \item{longitude}{longitude}
  \item{count}{Number of events at the location}
}}
\source{
\url{https://doi.org/10.6084/m9.figshare.4516772}
}
\usage{
data_geo_marker_plot
}
\description{
A (long-formatted) dataset designed to demonstrate different options for geo marker plots.
The dataset has been based on an academic dataset that will be published here https://doi.org/10.6084/m9.figshare.4516772
}
\keyword{datasets}
