% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_scatter_plot}
\alias{data_scatter_plot}
\title{Scatter plot data}
\format{A data frame with 2163 rows and 7 variables:
\describe{
  \item{chromosome}{Chromosome and genomic position of gene}
  \item{type}{Concordant (both sites are identical) or Discordant (mutations only in one replicate)}
  \item{x}{Allelic frequencies of gene in replicate 1}
  \item{y}{Allelic frequencies of gene in replicate 2}
}}
\source{
\url{https://dx.doi.org/10.6084/m9.figshare.4555441.v1}
}
\usage{
data_scatter_plot
}
\description{
A (long-formatted) dataset designed to demonstrate different options for scatter plot
}
\keyword{datasets}
