% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_stacked_bar_chart}
\alias{data_stacked_bar_chart}
\title{Stacked bar chart data}
\format{A data frame with 10 rows and 6 variables:
\describe{
  \item{Country}{price, in US dollars}
  \item{Business}{weight of the diamond, in carats}
  ...
}}
\source{
\url{http://www.diamondse.info/}
}
\usage{
data_stacked_bar_chart
}
\description{
A (wide-formatted) dataset designed to demonstrate different options for stacked bar charts.
}
\keyword{datasets}

