% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_world_shapefiles}
\alias{data_world_shapefiles}
\title{World Shapefiles}
\format{A "sf data.frame" with 241 elements, with 9 variables and 1 column containing the geographic feature corresponding to the observation.
\describe{
  \item{name}{Abbreviated Name of the country, i.e. N. Mariana Is. vs Northern Mariana Islands}
  \item{name_long}{Full name of the country, i.e. N. Mariana Is. vs Northern Mariana Islands}
  \item{type}{Type of country, i.e. Sovereign country, Dependency, etc}
  \item{continent}{Continent country belongs to}
  \item{region_un}{Region assignment from the UN}
  \item{subregion}{Subregion country belongs to, as per UN}
  \item{sovereignt}{Full name of country holding sovereignty over the country}
  \item{subunit}{UN subunit the country belongs to}
  \item{postal}{Postal shortcode for country}
}}
\source{
\url{http://www.diamondse.info/}
}
\usage{
data_world_shapefiles
}
\description{
A "sf data.frame" sourced from http://www.naturalearthdata.com/downloads/50m-cultural-vectors/ which includes shapefiles for all countries, with a scale of 1:50m.
}
\details{
If you are unfamiliar with using the "sf data.frame" object, please refer to vignette("GIS-features", package = "oidnChaRts")
}
\keyword{datasets}
