% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charts.R
\name{hc_stacked_bar_chart}
\alias{hc_stacked_bar_chart}
\title{highchart bar chart}
\usage{
hc_stacked_bar_chart(data = NA, categories_column = NA,
  measure_columns = NA, stacking_type = NA, ordering_function = c,
  explicit_order = NA)
}
\description{
highchart bar chart
}

