% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo-choropleth-chart.R
\name{geo_choropleth_chart}
\alias{geo_choropleth_chart}
\alias{leaflet_geo_choropleth_chart}
\title{geo_choropleth_chart}
\usage{
geo_choropleth_chart(spdf = NA, library = "leaflet",
  line.color = "#2c7bb6", line.popup = NULL, line.mouseover = NULL,
  start.color = "#fdae61", end.color = "#d7191c", both.color = "#7570b3",
  line.options = list(weight = 4), termini.legend = TRUE,
  termini.options = list(termini.radius = 4))

leaflet_geo_choropleth_chart(...)
}
\arguments{
\item{spdf}{A SpatialPolygonsDataFrame}

\item{...}{all arguments other than \code{data} and \code{library} provided to \code{geo_choropleth_chart}.}
}
\description{
\code{geo_choropleth_chart} creates a map comprised of shaded regions (where colour is used to allow visual comparison betweem region) using the specified library, which can be used in the library's \%>\% workflow. Data must be provided in long format.

\code{leaflet_geo_choropleth_chart} should not be used directly, it generates a map with great circles between points using Leaflet.
}
