% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter-plot.R
\name{scatter_plot}
\alias{scatter_plot}
\alias{hc_scatter_plot}
\alias{plotly_scatter_plot}
\title{Scatter plot.}
\usage{
scatter_plot(data = NA, library = "highcharter", x.column = ~x,
  y.column = ~y, color.column = ~color, marker.size = 5, traces.column,
  fillOpacity = 0.66)

hc_scatter_plot(...)

plotly_scatter_plot(...)
}
\arguments{
\item{data}{A dataframe, must be long-formatted.}

\item{library}{Which library to use, highchart is default.}

\item{x.column}{Column containing x-coordinates for data points}

\item{y.column}{Column containing y-cordinates for data points}

\item{color.column}{Column containing trace colour, default to "color"}

\item{marker.size}{Size of markers (circles, by default)}

\item{traces.column}{Column containing the traces/trace names}

\item{fillOpacity}{Opacity of markers, default 0.66}

\item{...}{all arguments provided to \code{scatter_plot}.}

\item{...}{all arguments provided to \code{scatter_plot}.}
}
\description{
\code{scatter_plot} creates a scatter plot using the specified library, which can be used in the library's \%>\% workflow. Data must be provided in long format.

\code{hc_scatter_plot} should not be used directly, it generates a scatter plot using highcharter.

\code{plotly_scatter_plot} should not be used directly, it generates a scatter plot using highcharter.
}
