% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charts.R
\name{stacked_bar_chart}
\alias{stacked_bar_chart}
\title{Stacked Bar Chart.}
\usage{
stacked_bar_chart(data = NA, library = "highcharter",
  categories.column = NA, categories.order = NA,
  subcategories.column = NA, value.column = NA, stacking.type = NA,
  subcategories.order = NA)
}
\arguments{
\item{data}{A dataframe, must be long-formatted.}

\item{library}{Which library to use, highchart is default.}

\item{categories.column}{Column containing the bar groupings (or categories), in a horizontally orientated barchart these will be the y-axis labels. Must be given as formula, i.e. ~country}

\item{categories.order}{Order for categories}

\item{subcategories.column}{Column containing different measures, i.e sub-categorisations. Must be given as formula, i.e. ~country}

\item{value.column}{Column containing bar values, i.e. lengths. Must be given as formula, i.e. ~country}

\item{stacking.type}{Stacking to apply to bars, default NULL. Possible values, c("percent", "normal")}

\item{subcategories.order}{Order for sub-categories, note for plotly the legend is ordered correctly but bars are reverse ordered}
}
\value{
Stacked bar chart of the class specified by the specified library.
}
\description{
\code{stacked_bar_chart} creates a stacked bar chart using the specified library, which can be used in the library's \%>\% workflow. Data must be provided in long format.
}
\section{Warning}{
 These should be considered utility functions only, the magic is contained within the htmlwidget library,
 it is perfectly possible breaking changes to the library and/or the underling JavaScript library will
 break these. In such case, refer to http://ox-it.github.io/OxfordIDN_htmlwidgets for generic tutorials on
the widely used htmlwidget libraries and how to select an alternative visualisation library.
}
\examples{
\dontrun{
stacked_bar_chart(1, 1)
stacked_bar_chart(1, "foo")
}

}

